+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/util/TestGenSampleFn.kt
@@ -1,14 +1,14 @@
-package org.cryptobiotic.rlauxe.core
+package org.cryptobiotic.rlauxe.util
 
+import org.cryptobiotic.rlauxe.core.*
 import org.cryptobiotic.rlauxe.doubleIsClose
 import org.cryptobiotic.rlauxe.doublePrecision
 import org.cryptobiotic.rlauxe.makeStandardComparisonAssorter
-import org.cryptobiotic.rlauxe.util.*
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
-class TestSamples {
+class TestGenSampleFn {
 
     @Test
     fun testSampleMeans() {
